feat(client): Migrate API calls and types to the new `api-v2` structure, deprecating `api` and `api-new` modules.
