+++ client/app/(landing)/search/page.tsx
@@ -1,6 +1,5 @@
 import VerticalArticleCard from "@/components/features/dashboard/VerticalArticleCard";
-import { getArticlesList } from "@/lib/api/news/service";
-import { ArticlesListResponse } from "@/lib/api/news/types";
+import { getArticlesList, ArticleListResponse } from "@/lib/api-v2";
 import { Categories, Countries } from "@/app/data";
 
 type Props = {
@@ -22,9 +21,9 @@ export default async function SearchPage({ searchParams }: Props) {
         country: country !== "all" ? country : undefined,
         category: category !== "all" ? category : undefined,
         limit: 20
-    }) as ArticlesListResponse;
+    });
 
-    const filteredArticles = response?.data || [];
+    const filteredArticles = response?.data?.data || [];
 
     // Helper lookup for labels
     const getLabel = (list: any[], val: string) =>
@@ -60,7 +59,7 @@ export default async function SearchPage({ searchParams }: Props) {
                                 year: 'numeric'
                             })}
                             views={`${article.views_count || 0} views`}
-                            imageSrc={article.image || article.image_url || "/images/placeholder.png"}
+                            imageSrc={article.image || "/images/placeholder.png"}
                         />
                     ))}
                 </div>
