+++ client/components/features/dashboard/LatestPostsSection.tsx
@@ -3,7 +3,7 @@
 import React, { useState } from 'react';
 import Link from 'next/link';
 import Image from 'next/image';
-import { ArticleResource } from "@/lib/api-new/types";
+import { ArticleResource } from "@/lib/api-v2";
 
 interface LatestPostsSectionProps {
     articles: ArticleResource[];
