+++ client/lib/api-v2/public/services/article/getArticleById.ts
@@ -4,9 +4,7 @@ import AXIOS_INSTANCE_PUBLIC from "../../axios-instance";
 import type { ArticleResource } from "../../../types/ArticleResource";
 import type { AxiosResponse } from "axios";
 
-export type ArticleByIdResponse = {
-  data: ArticleResource;
-}
+export type ArticleByIdResponse = ArticleResource;
 
 export async function getArticleById(id: string): Promise<AxiosResponse<ArticleByIdResponse>> {
   return AXIOS_INSTANCE_PUBLIC.get<ArticleByIdResponse>(`/articles/${id}`);
