--- client/lib/api/auth/index.ts
@@ -1,1 +1,0 @@
-export { useAuth } from "./store";
