--- client/lib/api/index.ts
@@ -1,2 +1,0 @@
-export * from "./news/service";
-export * from "./news/types";
