--- client/lib/api/news/types/index.ts
@@ -1,2 +1,0 @@
-export * from "./model";
-export * from "./response";
